+++ address_gateway.go
@@ -2,7 +2,6 @@ package braintree
 
 import (
 	"context"
-	"encoding/xml"
 )
 
 type AddressGateway struct {
@@ -10,13 +9,8 @@ type AddressGateway struct {
 }
 
 // Create creates a new address for the specified customer id.
-func (g *AddressGateway) Create(ctx context.Context, a *Address) (*Address, error) {
-	// Copy address so that field sanitation won't affect original
-	var cp Address = *a
-	cp.CustomerId = ""
-	cp.XMLName = xml.Name{Local: "address"}
-
-	resp, err := g.execute(ctx, "POST", "customers/"+a.CustomerId+"/addresses", &cp)
+func (g *AddressGateway) Create(ctx context.Context, customerID string, a *AddressRequest) (*Address, error) {
+	resp, err := g.execute(ctx, "POST", "customers/"+customerID+"/addresses", &a)
 	if err != nil {
 		return nil, err
 	}