+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Unreleased
+
+BACKWARDS INCOMPATIBILITES:
+
+* Drop support for versions of TLS prior to TLS 1.2.
+
+IMPROVEMENTS:
+
+* Add SkipAdvancedFraudChecking to TransactionOptions.
+
 ## 0.17.0 (January 5th, 2018)
 
 IMPROVEMENTS: